requirements.txt
# requirements.txt
python-telegram-bot[webhooks]==20.7
openai
supabase==2.5.0
gotrue==2.5.0
python-dotenv
pydub
requests
gspread==5.12.4
oauth2client==4.1.3
pymorphy3==1.2.1